Extract AbstractProperty.FinalizationState as top level class
